Firmware v1.4 for C64 Kernal switch
Version numbers aligned between firmware for different Kernal switches as functionality is the pretty much same for all switches.
Improved handling of restore key functionality while in menu. This was originally done for the C128 Kernal switch but was ported to the C64 firmware to keep them as similar as possible.
Code cleanup and some optimizations of main loop.
Thanks Pär Högberg!